Added login required to all views necessary

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -75,10 +76,16 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """
     Add a product to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -90,7 +97,7 @@ def add_product(request):
                 Please ensure the form is valid')
     else:
         form = ProductForm()
-        
+
     template = 'products/add_product.html'
     context = {
         'form': form,
@@ -99,10 +106,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product from the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -111,7 +123,8 @@ def edit_product(request, product_id):
             messages.success(request, 'Succesfully updated product!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Invalid entry, please ensure your form is valid!')
+            messages.error(request, 'Invalid entry, \
+                please ensure your form is valid!')
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.name}')
@@ -125,8 +138,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a porduct from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Succesfully deleted product!')
